data-importing/azurefunctions/requirements.txt
azure-storage
beautifulsoup4
requests